be/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1

# ── Config & Validation ───────────────────────────────────────────────────────
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0

# ── Auth & Security ───────────────────────────────────────────────────────────
bcrypt>=4.1.2
PyJWT>=2.8.0

# ── Cache & Rate Limit ────────────────────────────────────────────────────────
redis[asyncio]>=5.0.3

# ── Background Tasks ──────────────────────────────────────────────────────────
celery[redis]>=5.3.6

# ── File Upload ───────────────────────────────────────────────────────────────
python-multipart>=0.0.9
cloudinary>=1.39.0

# ── AI / Vision ───────────────────────────────────────────────────────────────
google-cloud-vision>=3.7.0

# ── Utilities ─────────────────────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
